requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
greenlet==1.1.0
idna==3.2
Markdown==3.3.4
markdown2==2.4.0
multidict==5.1.0
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.4.1
SQLAlchemy==1.4.18
typing-extensions==3.10.0.0
yarl==1.6.3
